@@ -29,19 +29,31 @@
             "name": "Sebastian De Deyne",
             "email": "sebastian@spatie.be",
             "homepage": "https://sebastiandedeyne.com"
+        },
+        {
+            "name": "Hanlin Wang",
+            "email": "605231181@qq.com",
+            "homepage": "https://wanghanlin.com"
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/events": "5.3.* || 5.4.* || 5.5.* || 5.6.*",
-        "illuminate/notifications": "5.3.* || 5.4.* || 5.5.* || 5.6.*",
-        "illuminate/queue": "5.3.* || 5.4.* || 5.5.* || 5.6.*",
-        "illuminate/support": "5.3.* || 5.4.* || 5.5.* || 5.6.*",
-        "pusher/pusher-php-server": "2.6.*"
+        "php": ">=7.0",
+        "illuminate/events": "5.5.* || 5.6.*",
+        "illuminate/notifications": "5.5.* || 5.6.*",
+        "illuminate/queue": "5.5.* || 5.6.*",
+        "illuminate/support": "5.5.* || 5.6.*",
+        "pusher/pusher-push-notifications": "^0.10"
     },
     "require-dev": {
-        "mockery/mockery": "^0.9.5",
-        "phpunit/phpunit": "4.*"
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "7.*"
+    },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "NotificationChannels\\PusherPushNotifications\\PusherPushNotificationsServiceProvider"
+            ]
+        }
     },
     "autoload": {
         "psr-4": {
