@@ -5,9 +5,9 @@ namespace NotificationChannels\PusherPushNotifications\Test;
 use Illuminate\Support\Arr;
 use NotificationChannels\PusherPushNotifications\Exceptions\CouldNotCreateMessage;
 use NotificationChannels\PusherPushNotifications\PusherMessage;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class MessageTest extends PHPUnit_Framework_TestCase
+class MessageTest extends TestCase
 {
     /** @var \NotificationChannels\PusherPushNotifications\PusherMessage */
     protected $message;
@@ -39,7 +39,7 @@ class MessageTest extends PHPUnit_Framework_TestCase
     public function by_default_it_will_send_a_message_to_ios()
     {
         $this->assertTrue(Arr::has($this->message->toArray(), 'apns'));
-        $this->assertFalse(Arr::has($this->message->toArray(), 'gcm'));
+        $this->assertFalse(Arr::has($this->message->toArray(), 'fcm'));
     }
 
     /** @test */
@@ -48,11 +48,11 @@ class MessageTest extends PHPUnit_Framework_TestCase
         $this->message->ios();
 
         $this->assertTrue(Arr::has($this->message->toArray(), 'apns'));
-        $this->assertFalse(Arr::has($this->message->toArray(), 'gcm'));
+        $this->assertFalse(Arr::has($this->message->toArray(), 'fcm'));
 
         $this->message->android();
 
-        $this->assertTrue(Arr::has($this->message->toArray(), 'gcm'));
+        $this->assertTrue(Arr::has($this->message->toArray(), 'fcm'));
         $this->assertFalse(Arr::has($this->message->toArray(), 'apns'));
     }
 
@@ -69,7 +69,7 @@ class MessageTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals('myTitle', Arr::get($this->message->toiOS(), 'apns.aps.alert.title'));
 
-        $this->assertEquals('myTitle', Arr::get($this->message->toAndroid(), 'gcm.notification.title'));
+        $this->assertEquals('myTitle', Arr::get($this->message->toAndroid(), 'fcm.notification.title'));
     }
 
     /** @test */
@@ -79,7 +79,7 @@ class MessageTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals('myBody', Arr::get($this->message->toiOS(), 'apns.aps.alert.body'));
 
-        $this->assertEquals('myBody', Arr::get($this->message->toAndroid(), 'gcm.notification.body'));
+        $this->assertEquals('myBody', Arr::get($this->message->toAndroid(), 'fcm.notification.body'));
     }
 
     /** @test */
@@ -89,7 +89,7 @@ class MessageTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals('mySound', Arr::get($this->message->toiOS(), 'apns.aps.sound'));
 
-        $this->assertEquals('mySound', Arr::get($this->message->toAndroid(), 'gcm.notification.sound'));
+        $this->assertEquals('mySound', Arr::get($this->message->toAndroid(), 'fcm.notification.sound'));
     }
 
     /** @test */
@@ -105,13 +105,13 @@ class MessageTest extends PHPUnit_Framework_TestCase
     {
         $this->message->icon('myIcon');
 
-        $this->assertEquals('myIcon', Arr::get($this->message->toAndroid(), 'gcm.notification.icon'));
+        $this->assertEquals('myIcon', Arr::get($this->message->toAndroid(), 'fcm.notification.icon'));
     }
 
     /** @test */
     public function it_will_throw_an_exception_when_an_unsupported_platform_is_used()
     {
-        $this->setExpectedException(CouldNotCreateMessage::class);
+        $this->expectException(CouldNotCreateMessage::class);
 
         $this->message->platform('bla bla');
     }
@@ -122,6 +122,6 @@ class MessageTest extends PHPUnit_Framework_TestCase
         $this->message->ios()->withAndroid(new PusherMessage());
 
         $this->assertTrue(Arr::has($this->message->toArray(), 'apns'));
-        $this->assertTrue(Arr::has($this->message->toArray(), 'gcm'));
+        $this->assertTrue(Arr::has($this->message->toArray(), 'fcm'));
     }
 }
